Refactor of monster data, better show/hiding of text
Monster parts are now grouped by an id for easy visual customization in the skin on a per-group basis.

Removable parts are now contained in the same collection as parts. Parts are now sorted in the skin so removables appear first.

Parts and monster status effects can now be filtered by group id in the main config using regex.

Some changes to config options on how text gets displayed between numbers and percents.